Fix php 32 bit errors

+++ src/BackupDestination/Backup.php
@@ -55,7 +55,7 @@ class Backup
     /**
      * Get the size in bytes.
      */
-    public function size(): int
+    public function size(): float
     {
         if ($this->size === null) {
             if (! $this->exists()) {
